BF: Avoid crash when hitting a file that was already moved for BIDS mapping
This is now necessary, because we are calling the BIDS mapping code
after each data record ZIP file was add-extracted, and prior any `datalad
save` (instead of once at the end all processing). This can lead to the
situation that a file from the same instance with already moved, but
`status()` still reports its original location.

Alternatively we could figure out which files were added from the last
ZIP, and avoid the `status()` call completely. But keeping the original
approach for now (for simplicity of the diff) until we see evidence that
more substantial changes are necessary.