Maintain BIDS and native dataset layout in parallel
The decision whether a BIDS representation is desired, is now made via
`init`. This seems to make more sense than having to indicate that
for each call to `update`.

The ability to customize the naming of a directory for all unrecognized
content in the BIDS layout has been removed. I consider that a needless
complication with the switch to multiple branches.

Test are expanded to cover layout differences.

Fixes gh-38