Inherit gitattributes in incoming branches
Because we build the branches from scratch (i.e. not derived from
master), we are missing the standard .gitattributes file. Consequently,
dataset updates are susceptible to annex backend configuration changes
on a workstation, because the dataset incoming branch doesn't specify
any.

With this change, a .gitattributes file is inherited from the current
state in master on each update, of any incoming branch.

Fixes gh-54