Fix update --drop extracted crash (missing archives SR)
We now check whether there is a `datalad-archives` special remote,
before attempting to drop from it. We could have used an approach
that check if there is candidate content for this SR, but that would
likely duplicate logic and be more fragile wrt future changes.

Here we test for the error/crash condition directly.

Debugged with @loj

Fixes gh-69