BF: introduce helper to ensure newline characters in .gitattributes files
Within datalad core, datalad/datalad@fa96e6a introduced a newline to .gitattributes
largefile configurations written by datalad at repo initialization. This change has
been a part of the code base since version 0.15, if I am not mistaken.
The presence of the newline causes a mismatch between what should be identical
.gitattributes files across two branches, ultimately causing a merge conflict and
resulting in https://github.com/datalad/datalad-ukbiobank/issues/82.
This is because the .gitattributes file of one branch is written using pathlibs
write_text method with input from a cat-file -p command on the main branch.
As write_text() does not support addition of a newline character until Python 3.10,
the most feasible way to ensure the presence of the newline is to append it if it
isn't there via string operations. In this change, this is done with the introduction
of a helper function. As this code base does not have any utils.py and the helper is
small, I added it directly into init.py - if there are more helpers introduced, it
would probably make sense to create a utils.py for them eventually.